Added observers to classes, and made VBO template safer
Added observers to the GL wrappers for Shader, VBO, IBO and VAO.
Additionally, instead of VBO's constructor and updater being templated,
made the VBO class templated on a single type, so that the underlying
vector template cannot be changed later. The VBO is templated on a
single type so that constructor and updater take same template
parameter.